+++ mobile/app/(tabs)/_layout.tsx
@@ -3,8 +3,11 @@ import { Tabs } from "expo-router";
 import { StyleSheet } from "react-native";
 import MaterialIcons from "@expo/vector-icons/MaterialIcons";
 import FontAwesome from "@expo/vector-icons/FontAwesome";
+import { useAuth } from "@clerk/clerk-expo";
 
 export default function TabLayout() {
+  const { isSignedIn } = useAuth();
+
   return (
     <Tabs
       screenOptions={{
@@ -29,6 +32,14 @@ export default function TabLayout() {
           tabBarIcon: ({ color }) => (
             <FontAwesome name="cog" size={24} color={color} />
           ),
+          href: isSignedIn ? "/settings" : null,
+        }}
+      />
+      <Tabs.Screen
+        name="login"
+        options={{
+          title: "Login",
+          href: null,
         }}
       />
     </Tabs>
