+++ mobile/app/(tabs)/settings.tsx
@@ -2,16 +2,11 @@ import React from "react";
 import { View, Text, StyleSheet } from "react-native";
 import { Link } from "expo-router";
 import { A } from "@expo/html-elements";
-import { useAuth } from "@clerk/clerk-expo";
 
 export default function Settings() {
-  const { userId, isSignedIn } = useAuth();
   return (
     <View style={styles.container}>
       <Text style={styles.title}>Settings</Text>
-      <Text style={styles.description}>
-        {isSignedIn ? `Signed in as ${userId}` : "Not signed in"}
-      </Text>
       <Text style={styles.description}>
         XTrack was created out of a need to better track finances, by{" "}
         <A href="https://bamboostud.io">Team Bamboo 🎋</A>.
@@ -21,7 +16,7 @@ export default function Settings() {
       </Link>
     </View>
   );
-};
+}
 
 const styles = StyleSheet.create({
   container: {
